refactor: 훅과 로직 분리
